requirements.txt
psycopg2==2.9.7
python-dotenv==1.0.0
Scrapy==2.10.0
